# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Usuarios",
    "version": "1.0.0",
    "description": "Documentación de la API para gestionar usuarios"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Error en la solicitud"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Obtener un usuario por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar un usuario existente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "400": {
            "description": "Error en la solicitud"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado correctamente"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "400": {
            "description": "Error en la solicitud"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string",
            "format": "email"
          },
          "contrasena": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "fotoPerfil": {
            "type": "string",
            "format": "uri"
          },
          "peso": {
            "type": "number",
            "nullable": true
          },
          "altura": {
            "type": "number",
            "nullable": true
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date"
          },
          "genero": {
            "type": "string"
          }
        },
        "required": ["correo", "contrasena", "nombre", "fechaNacimiento", "genero"]
      }
    }
  }
}